Fix #1100: use BigInteger

+++ chemclipse/plugins/org.eclipse.chemclipse.ux.extension.xxd.ui/src/org/eclipse/chemclipse/ux/extension/xxd/ui/support/charts/ScanDataSupport.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017, 2021 Lablicate GmbH.
+ * Copyright (c) 2017, 2022 Lablicate GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -8,9 +8,11 @@
  * 
  * Contributors:
  * Dr. Philip Wenig - initial API and implementation
+ * Lorenz Gerber - fix integer overflow for signal label
  *******************************************************************************/
 package org.eclipse.chemclipse.ux.extension.xxd.ui.support.charts;
 
+import java.math.BigDecimal;
 import java.text.DecimalFormat;
 
 import org.eclipse.chemclipse.model.core.IChromatogram;
@@ -89,7 +91,7 @@ public class ScanDataSupport {
 			//
 			builder.append(" | ");
 			builder.append("Signal: ");
-			builder.append((int)scan.getTotalSignal());
+			builder.append(BigDecimal.valueOf(scan.getTotalSignal()).toBigInteger());
 			//
 			if(scan instanceof IScanMSD) {
 				IScanMSD scanMSD = (IScanMSD)scan;
